pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rdc-bigquery-bridge"
version = "0.1.0"
description = "Real-Time RDC-to-BigQuery Data Bridge"
authors = ["Alexander Morosow <alex@refikanadolstudio.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.0"
google-cloud-bigquery-storage = "^2.16.0"
google-cloud-bigquery = "^3.11.0"
pyyaml = "^6.0"
tenacity = "^8.2.0"
protobuf = "^4.24.0"
asyncio-mqtt = "^0.16.1"
msgpack = "^1.1.1"
pyarrow = "^22.0.0"
fastavro = "^1.9.0"
pandas = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
rdc-bridge = "src.main:cli_main"
rdc-export = "src.exporter.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
